Split peer_posix_screen.c into two
Part of this functionality is needed by the unit tests and the
posix tools (saimport,salink,sexport) and part is just needed
by the tools.  We don't want this part in the unit tests.
