+++ src/peer.h
@@ -21,10 +21,10 @@
 
 #include "error.h"
 
-#ifdef SPECTRUM
-#include "peer_zx.h"
-#elif defined(UNITTESTS)
+#ifdef UNITTESTS
 #include "peer_unit.h"
+#elif defined(SPECTRUM)
+#include "peer_zx.h"
 #endif
 
 void specasm_peer_write_state_e(const char *fname, uint16_t checksum);
@@ -36,6 +36,11 @@ void specasm_text_set_flash(uint8_t x, uint8_t y, uint8_t attr);
 void specasm_text_printch(char ch, uint8_t x, uint8_t y, uint8_t attr);
 uint8_t specasm_text_print(const char *str, uint8_t x, uint8_t y, uint8_t attr);
 void specasm_text_clear(uint8_t x, uint8_t y, uint8_t l, uint8_t attr);
+
+#ifndef SPECTRUM
+int itoa(int n, char *s, unsigned char radix);
+int utoa(int n, char *s, unsigned char radix);
+#endif
 
 /*
  * Flushes any buffered output and sets the text cursor to
